Investigating build error

+++ src/doddle/words.py
@@ -157,7 +157,7 @@ class WordSeries:
 
         is_slice = isinstance(s, slice)
         is_mask = isinstance(s, np.ndarray) and s.dtype == np.bool8
-        is_index = isinstance(s, np.ndarray) and s.dtype == np.int32
+        is_index = isinstance(s, np.ndarray) and s.dtype == np.int_
         can_index = is_slice or is_mask or is_index
 
         if can_index:
